requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
supabase==1.2.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
flask-login==0.6.3 
